# doc/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Spika for Business API v3
  description: >-
    New set of API to use Spika for Business. First of all, you must get an
    access token by calling sign in API. 
schemes:
  - http
  - https
host: 'spika.chat'
basePath: /api/v3
...
paths:
...
  /messages:
    post:
      tags:
        - Messages2
      summary: Send message
      description: 'Send message to user,group and room.'
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - name: body
          required: true
          in: body
          description: Parameters to send
          schema:
            properties:
              targetType:
                type: string
                description: 1.private 2.group 3.room
              target:
                type: string
              messageType:
                type: string
                description: 1.text 2.file 3.location 4.contact 5.sticker
              message:
                type: string
                description: required if messageType is 1
              file:
                $ref: '#/definitions/File'
              thumb:
                $ref: '#/definitions/File'
      responses:
        '200':
          description: The message sent
          schema:
            properties:
              message:
                $ref: '#/definitions/Message'
              user:
                $ref: '#/definitions/User'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
  '/messages/{messageId}':
    put:
      tags:
        - Messages2
      summary: Edit mesasge details
      description: 'Edit message of user, group and room.'
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          type: string
          name: messageId
          description: Message id to update message
          required: true
        - name: body
          in: body
          description: Parameters to send
          required: true
          schema:
            properties:
              message:
                type: string
                description: required if messageType is 1
      responses:
        '200':
          description: Edited message
          schema:
            properties:
              message:
                $ref: '#/definitions/Message'
              user:
                $ref: '#/definitions/User'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
    delete:
      tags:
        - Messages2
      summary: Delete group
      description: Delete group corresonding to group id.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: messageId
          description: mesasgeId to delete message
          type: string
          required: true
      responses:
        '200':
          description: Deleted message
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
...
  '/groups/{groupId}/messages/':
    get:
      tags:
        - Groups
      summary: Get list of messages sent to group
      description: Get list of messages sent to group.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: groupId
          description: group ID to get group list
          type: string
          required: true
        - in: query
          name: offset
          description: The offset number of get messages
          type: integer
          format: int32
        - in: query
          name: limit
          description: The limit number of get messages
          type: integer
          format: int32
        - in: query
          name: sort
          description: >-
            Sort by fields set by string like [field name]:asc, [field
            name]:desc
          type: string
        - in: query
          name: fields
          description: >-
            Use for filtering params. Only return set fields if you set them,
            otherwise all fields will be returned.
          type: array
          items:
            type: string
      responses:
        '200':
          description: List of messages sent to group
          schema:
            properties:
              messages:
                type: array
                items:
                  $ref: '#/definitions/Message'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
...
  '/rooms/{roomId}/messages/':
    get:
      tags:
        - Rooms
      summary: Get list of messages sent to room
      description: Get list of messages sent to room.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room list
          type: string
          required: true
        - in: query
          name: offset
          description: The offset number of get messages
          type: integer
          format: int32
        - in: query
          name: limit
          description: The limit number of get messages
          type: integer
          format: int32
        - in: query
          name: sort
          description: >-
            Sort by fields set by string like [field name]:asc, [field
            name]:desc
          type: string
        - in: query
          name: fields
          description: >-
            Use for filtering params. Only return set fields if you set them,
            otherwise all fields will be returned.
          type: array
          items:
            type: string
      responses:
        '200':
          description: list of messages sent to toom
          schema:
            properties:
              messages:
                type: array
                items:
                  $ref: '#/definitions/Message'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
  '/private/{targetUserId}/messages/':
    get:
      tags:
        - Private Chat
      summary: Get list of messages sent to private chat
      description: Get list of messages sent to private chat.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: targetUserId
          description: Target user ID to get private chat list
          type: string
          required: true
        - in: query
          name: offset
          description: The offset number of get messages
          type: integer
          format: int32
        - in: query
          name: limit
          description: The limit number of get messages
          type: integer
          format: int32
        - in: query
          name: sort
          description: >-
            Sort by fields set by string like [field name]:asc, [field
            name]:desc
          type: string
        - in: query
          name: fields
          description: >-
            Use for filtering params. Only return set fields if you set them,
            otherwise all fields will be returned.
          type: array
          items:
            type: string
      responses:
        '200':
          description: list of messages sent to private chat
          schema:
            properties:
              messages:
                type: array
                items:
                  $ref: '#/definitions/Message'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
...
securityDefinitions:
  api_key:
    type: apiKey
    name: apiKey
    in: header
definitions:
...
  User:
    properties:
      id:
        type: string
      name:
        type: string
      sortName:
        type: string
      description:
        type: string
      userid:
        type: string
      password:
        type: string
      created:
        type: number
        format: double
      token:
        type: array
        items:
          type: string
      pushToken:
        type: array
        items:
          type: string
      voipPushToken:
        type: array
        items:
          type: string
      organizationId:
        type: object
        properties:
          type:
            type: string
          index:
            type: boolean
            example: true
      avatar:
        $ref: '#/definitions/Avatar'
      groups:
        $ref: '#/definitions/GroupList'
      permission:
        type: integer
        format: int32
      muted:
        type: array
        items:
          type: string
      blocked:
        type: array
        items:
          type: string
      devices:
        type: array
        items:
          type: string
      UUID:
        type: array
        items:
          type: string
      phoneNumber:
        type: string
      activationCode:
        type: string
...
  File:
    properties:
      size:
        type: integer
        format: int32
      fileId:
        type: string
      mimeType:
        type: string
...
  GroupList:
    type: array
    items:
      type: string
...
  Message:
    properties:
      _id:
        type: string
      message:
        type: string
      roomID:
        type: string
      created:
        type: number
        format: double
      user:
        $ref: '#/definitions/User'
      file:
        properties:
          file:
            $ref: '#/definitions/File'
          thumb:
            $ref: '#/definitions/File'
...
  Avatar:
    properties:
      picture:
        type: object
        properties:
          originalName:
            type: string
          size:
            type: integer
            format: int32
          mimeType:
            type: string
          nameOnServer:
            type: string
      thumbnail:
        type: object
        properties:
          originalName:
            type: string
          size:
            type: integer
            format: int32
          mimeType:
            type: string
          nameOnServer:
            type: string
...